federated/Server/requirements.txt
numpy
python-decouple
scikit_learn
tensorflow
tqdm
